+++ bricks/basic-bricks/src/i18n/constants.ts
@@ -11,8 +11,9 @@ export enum K {
   BRICK_ERROR = "BRICK_ERROR",
   LEGACY_TEMPLATE_ERROR = "LEGACY_TEMPLATE_ERROR",
   HELP = "HELP",
-  CHINESE = "CHINESE",
-  ENGLISH = "ENGLISH",
+  CANCEL = "CANCEL",
+  OK = "OK",
+  COVERT_TO_LANGUAGE = "COVERT_TO_LANGUAGE",
 }
 
 export type Locale = { [key in K]: string };