+++ bricks/basic-providers/src/i18n/locales/zh.ts
@@ -1,7 +1,0 @@
-import { Locale, K } from "../constants";
-
-const locale: Locale = {
-  [K.BASIC_PROVIDERS]: "Basic Providers",
-};
-
-export default locale;