+++ bricks/next-builder/src/builder-container/EventStreamCanvas/buildBrickEventDownstreamTree.ts
@@ -1,5 +1,3 @@
-/* istanbul ignore file */
-// Todo(steve): Ignore tests temporarily for potential breaking change in the future.
 import {
   BrickEventHandler,
   MessageConf,