+++ bricks/next-builder/src/builder-container/EventStreamCanvas/buildBrickEventUpstreamTree.ts
@@ -1,5 +1,3 @@
-/* istanbul ignore file */
-// Todo(steve): Ignore tests temporarily for potential breaking change in the future.
 import { isEmpty } from "lodash";
 import { HierarchyPointNode } from "d3";
 import {